requirements.txt
pytorch>=0.4.1
soundfile
tensorboardX
